Fixed mount usage

+++ src/model/Map/Zone/TileRelatedClasses/River.java
@@ -35,7 +35,7 @@ public class River implements Saveable {
         double sumSpeed = Math.pow(sumX, 2) + Math.pow(sumY,2);
         int newEntitySpeed = (int) Math.sqrt(sumSpeed);
 
-        int degrees = (int) Math.toDegrees(Math.atan(sumY/((double)sumX)));
+        int degrees = (int) Math.toDegrees(Math.atan((double)sumY/((double)sumX)));
 
         entity.setCurrentSpeed(newEntitySpeed);
         return Direction.getDirectionClosest(degrees);
